requirements.txt
appdirs==1.4.4
asgiref==3.2.10
beautifulsoup4==4.9.3
boto3==1.15.15
botocore==1.18.15
cchardet==2.1.6
certifi==2020.6.20
chardet==3.0.4
conda==4.3.16
distlib==0.3.1
Django==3.1.2
djangorestframework==3.12.1
extract-social-media==0.4.0
filelock==3.0.12
future==0.18.2
html-to-etree==0.2.0
html5==0.0.9
htmlmetadata==1.1
idna==2.10
jmespath==0.10.0
lxml==4.5.2
metadata-parser==0.10.0
oauthlib==3.1.0
pipenv==2020.8.13
pycosat==0.6.3
python-dateutil==2.8.1
python-twitter==3.5
pytz==2020.1
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.3
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
urllib3==1.25.10
virtualenv==20.0.31
virtualenv-clone==0.5.4
w3lib==1.22.0
webencodings==0.5.1
